requirements.txt
pycurl==7.43.0
PyYAML==3.12